requirements.txt
Django==2.2.6
django-background-tasks==1.2.0
django-compat==1.0.15
mysqlclient==1.4.4
pyactiveresource==2.1.2
pytz==2019.3
PyYAML==5.1.2
ShopifyAPI==5.1.2
six==1.12.0
sqlparse==0.3.0
